third_party/blink/web_tests/fast/dom/shadow/shadow-boundary-crossing-expected.txt

CONSOLE WARNING: line 125: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests to ensure that shadow DOM boundary is not crossed during event propagation. Can only run within DRT.

See bug 46015 for details.

Mutation events should not propagate out of the shadow DOM: PASS
The selectstart event should not propagate out of the shadow DOM: PASS
The mouseover/mouseout event between two elements inside the same shadow subtree should not propagate out of the shadow DOM: PASS
The mouseover event in a shadow subtree, where related target is the tree host should not escape out of shadow DOM: PASS
Events with relatedTarget should not escape out of shadow subtree when its host is the target: PASS
The mouseover/mouseout event on a shadow subtree host should propagate out of the shadow DOM: PASS
Label should look beyond shadow boundary to detect if it encloses its associated element: PASS
Event's relatedTarget should be retargeted: PASS
Other events should be retargeted: PASS
After event dispatch, the event object should not reveal shadow DOM: PASS
Focusing same shadow DOM element repeatedly should not trigger multiple focus/blur events: PASS
